requirements.txt
anthropic
youtube-transcript-api
scrapetube
python-dotenv
